pyproject.toml
[project]
name = "singtown-ai-launcher"
version = "0.1.0"
description = "Launcher for Singtown AI application"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "filelock>=3.20.0",
    "pystray>=0.19.5",
    "pywebview>=6.1",
]

[dependency-groups]
dev = [
    "nuitka>=2.8.4",
]
